add: api de carpetas completada

+++ src/service/folders.service.js
@@ -1,5 +1,9 @@
 import mongoose from "mongoose";
 import folderModel from "../models/Folder.model";
+import {
+	deleteRelationsBy_idFolder,
+	findRelationsBy_FolderId,
+} from "./folderHeroRelation.service";
 
 export const newFolder_service = async (name, idUser) => {
 	try {
@@ -7,7 +11,7 @@ export const newFolder_service = async (name, idUser) => {
 
 		if (folderExists) return { error: { message: "la carpeta ya existe" } };
 
-		const folder = await folderModel.create(folder);
+		const folder = await folderModel.create({ name, idUser });
 
 		return folder;
 	} catch (error) {
@@ -17,7 +21,10 @@ export const newFolder_service = async (name, idUser) => {
 
 export const findByUserId_service = async (idUser) => {
 	try {
-		const folders = await folderModel.find({ idUser }).sort({ name: 1 });
+		const folders = await folderModel
+			.find({ idUser }, { idUser: false })
+			.sort({ name: 1 });
+
 		return folders;
 	} catch (error) {
 		console.log(error);
@@ -33,13 +40,13 @@ export const findById_service = async (id) => {
 	}
 };
 
-export const updateFolderName_service = async (idFolder, newName) => {
+export const updateFolderName_service = async (idFolder, name) => {
 	try {
-		const folder = await folderModel.findById(idFolder);
+		const folder = await folderModel.findById(idFolder, { idUser: false });
 
 		if (!folder) return { error: { message: "carpeta no existe" } };
 
-		folder.name = newName;
+		folder.name = name;
 
 		return await folder.save();
 	} catch (error) {
@@ -49,7 +56,13 @@ export const updateFolderName_service = async (idFolder, newName) => {
 
 export const delFolder_service = async (idFolder) => {
 	try {
-		// todo: eliminar sus relaciones con los heroes
+		const folder = await folderModel.findById(idFolder);
+
+		// todo: ver que no sea la carpeta favorito
+		if (!folder) return;
+		if (folder.name == "favorite") return;
+
+		await deleteRelationsBy_idFolder(idFolder);
 
 		const result = await folderModel.deleteOne({ _id: idFolder });
